(feat): confirm modal

+++ src/app/features/messages/messages.component.ts
@@ -8,6 +8,7 @@ import { RouterLink } from '@angular/router';
 import { JsonPipe, TitleCasePipe } from '@angular/common';
 import { TimeagoModule } from 'ngx-timeago';
 import { PaginationModule } from 'ngx-bootstrap/pagination';
+import { ConfirmService } from '../../core/services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -24,7 +25,7 @@ export class MessagesComponent implements OnInit{
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) {}
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) {}
 
   ngOnInit(): void {
     this.loadMessages();
@@ -40,8 +41,12 @@ export class MessagesComponent implements OnInit{
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
+      if(result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
+        })
+      }
     })
   }
